selenium docker rabbitmq/input/requirements.txt
requests
pika
Pillow
selenium
cairosvg
pymongo